directory: Don't mark moved file that is gone
With any operation (add, remove, move), it's possible (and likely)
to get multiple emmissions of the same event (one from nautilus
file changes queue and one from nautilus monitor).  nautilus-directory
handles this well for add and remove, as a file can only be removed
once and added files are always treated as newly added files.  For
moved files however, we check if the destination already exists, and
if it does, we mark it as gone.  This is problematic when the second
call to move is processed, so let's check if the "from_file" is null,
in order to determine if the move already happened.