monitor: Add ability to hint moved files
The file monitor will almost always pick up on a file move
before the nautilus file changes queue.  In the previous
commit, we now have the ability to detect moved evens which
is superior than a removed/created pair (which incorrectly
marks the nautilus-file as gone).  Still, we only get the
proper moved even if both the source and destination directories
are being monitored, which in a lot of cases isn't true.  In order
to more consistently move nautilus-files internally, allow
hinting when a file is moved (giving us either the missing source
or the missing destination).