files-view: Enable copy/paste action in fc save mode
Copy / paste should be allowed in fc save mode.  It is akin to the new folder
action.  It's common practice to have a directory hierarchy, and one might
want to copy paste that hierarchy instead of creating all of the new
folders by hand.  Another potential use case is to copy / paste a backup
copy of a file before overwriting a file.

This commit only enables the actions which would enable the keyboard shortcuts
but not the menu entries.